refactor(lint): cross-feature 내부 import 금지 (boundaries) + barrel 경유 정리 (P2-3)
- eslint-plugin-boundaries 도입: feature 는 다른 feature 를 barrel(index.ts) 로만
  import 하도록 boundaries/entry-point 로 강제 (같은 feature 내부/ app.module 은 비대상)
- user 서비스·모듈 6곳의 deep-path cross-feature import 를 barrel 경유로 정리:
  - user-{wishlist,recent-view}.service → @/features/product (ProductRepository)
  - user-{order,mypage}.service → @/features/order (OrderRepository)
  - user.module → @/features/{order,product} (Order/ProductModule)
- 위반 0건. 향후 cross-feature 내부 결합 회귀 방지.

NOTE: boundaries/entry-point 는 v6 에서 deprecated(→ boundaries/dependencies).
신규 API 가 default-disallow 라 #126 no-restricted 와 중복·충돌 소지가 있어
현 시점은 검증된 entry-point 유지. 마이그레이션은 후속 과제(Tech Debt)로 분리.
